% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossloadcheck.R
\name{crossloadcheck}
\alias{crossloadcheck}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
crossloadcheck(data = '',
                sigLevel = .05,
                scalingCrit = '',
                stepPrev = '')
}
\arguments{
\item{data}{The data matrix to be examined.}

\item{sigLevel}{The significance level threshold, default is .05.}

\item{scalingCrit}{The criterion used to select the scaling indicators, default is 'factorloading_R2.'}

\item{stepPrev}{The output object from the last step from the function stepN_EFAmiiv.}
}
\description{
This function checks for crossloadings when there are more than 2 problematic variables after creating a new factor.
}
\examples{
\dontrun{
crossloadcheckobj <- crossloadcheck(data = mydata, stepPrev = step1)
}
}
\author{
Lan Luo
}
