% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbadvar.R
\name{getbadvar}
\alias{getbadvar}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
getbadvar(fit = '',
                sigLevel = .05)
}
\arguments{
\item{fit}{The fit object obtained from miive function from the MIIVsem package.}

\item{sigLevel}{The significance level threshold, default is .05.}
}
\description{
This function extracts problematic variables from a model fit object.
}
\examples{
\dontrun{
mybadvar <- getbadvar(fit = mydata)
}
}
\author{
Lan Luo
}
