% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getbadvar_multi.R
\name{getbadvar_multi}
\alias{getbadvar_multi}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
getbadvar_multi(fit = '',
                sigLevel = .05,
                num_factor = '',
                varPerFac = '')
}
\arguments{
\item{fit}{The fit object obtained from miive function from the MIIVsem package.}

\item{sigLevel}{The significance level threshold, default is .05.}

\item{num_factor}{The number of factors, default is 1.}

\item{varPerFac}{It is a list storing the current variables on each factor.}
}
\description{
This function extracts problematic variables from a multi-factor model fit object by factor.
}
\examples{
\dontrun{
badvarfrommultifit <- getbadvar_multi(fit = mydata, varPerFac = myvarperfac)
}
}
\author{
Lan Luo
}
