% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsigsarganNnonsigloading.R
\name{getsigsarganNnonsigloading}
\alias{getsigsarganNnonsigloading}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
getsigsarganNnonsigloading(modellist = '',
                           fitlist = '',
                           sigLevel = .05)
}
\arguments{
\item{modellist}{A list of models with the same variables that only differ by the scaling indicator.}

\item{fitlist}{A list of miive fit objects based on the models in modellist.}

\item{sigLevel}{The significance level threshold, default is .05.}
}
\description{
This function extracts variables that have significant sargan statistics and/or non-significant factor loadings from a list of models.
}
\examples{
\dontrun{
myproblematicvariables <- getsigsarganNnonsigloading(modellist = mymodellist, fitlist = myfitlist)
}
}
\author{
Lan Luo
}
