% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_order.R
\name{r2_order}
\alias{r2_order}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
r2_order(data = '')
}
\arguments{
\item{data}{The data matrix.}
}
\description{
This function calculates the R2 values for each variable in the data when the rest variables are regressed on it.
}
\examples{
\dontrun{
r2values <- r2_order(data = mydata)
}
}
\author{
Lan Luo
}
