% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step1_EFAmiiv.R
\name{step1_EFAmiiv}
\alias{step1_EFAmiiv}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
step1_EFAmiiv(data='',
                 sigLevel = .05,
                 scalingCrit = "sargan+factorloading_R2",
                 correlatedErrors = NULL = '')
}
\arguments{
\item{data}{The data matrix.}

\item{sigLevel}{The significance level threshold, default is .05.}

\item{scalingCrit}{The criterion used to select the scaling indicators, default is 'factorloading_R2.'}

\item{correlatedErrors}{The pairs of variables whose errors should be correlated in the model search procedure, default is NULL.}
}
\description{
This function is the first step of EFAmiie and it runs a one factor model on the provided data.
}
\examples{
\dontrun{
myonefacefamiivobj <- step1_EFAmiiv(data = mydata, correlatedErrors = 'v3~~v4')
}
}
\author{
Lan Luo
}
