% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepN_EFAmiiv.R
\name{stepN_EFAmiiv}
\alias{stepN_EFAmiiv}
\title{Exploratory Factor Analysis Using Model Implied Instrumental Variables.}
\usage{
stepN_EFAmiiv(data='',
                 sigLevel = .05,
                 scalingCrit = "sargan+factorloading_R2",
                 stepPrev = '')
}
\arguments{
\item{data}{The data matrix to be examined.}

\item{sigLevel}{The significance level threshold, default is .05.}

\item{scalingCrit}{The criterion used to select the scaling indicators, default is 'factorloading_R2.'}

\item{stepPrev}{The output object from the last step from the function crossloadcheck.}
}
\description{
This function checks if a new factor is needed from the last step model, or if crossloadings should be checked.
}
\examples{
\dontrun{
myefamiivobj_stepN <- stepN_EFAmiiv(data = mydata, stepPrev = crossloadcheckobj)
}
}
\author{
Lan Luo
}
